requirements.txt
# Core Django and DRF
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Additional Utilities
Pillow==10.2.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1

# Optional Advanced Features
celery==5.3.6
redis==5.0.1
boto3==1.34.28

# Development and Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0